requirements.txt
PySide6
SQLAlchemy
ttkbootstrap
fpdf2
openpyxl
arabic-reshaper
python-bidi
matplotlib
pillow
Flask
